chore: extracting fonts into a separate folder
