+++ app/layout.tsx
@@ -1,27 +1,6 @@
 import type { Metadata } from "next";
-import { Roboto, Poppins, Fira_Sans } from "next/font/google";
+import { roboto, poppins } from "@/config/fonts";
 import "./globals.css";
-
-const roboto = Roboto({
-  subsets: ["latin"],
-  variable: "--font-roboto",
-  weight: ['400', '500', '700', '900'],
-  style: ['normal', 'italic'],
-});
-
-const poppins = Poppins({
-  subsets: ["latin"],
-  variable: "--font-poppins",
-  weight: ['400', '500', '700', '900'],
-  style: ['normal', 'italic'],
-});
-
-const firaSans = Fira_Sans({
-  subsets: ["latin"],
-  variable: "--font-fira-sans",
-  weight: ['400', '500', '700', '900'],
-  style: ['normal', 'italic'],
-});
 
 export const metadata: Metadata = {
   title: "Help next door",
@@ -35,7 +14,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={`${roboto.variable} ${poppins.variable} ${firaSans.variable}`}>{children}</body>
+      <body className={`${roboto.variable} ${poppins.variable}`}>{children}</body>
     </html>
   );
 }
