+++ app/page.tsx
@@ -1,7 +1,9 @@
 export default function HomePage() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between">
-      <h1>Welcome to the start</h1>
+    <main className="flex min-h-screen flex-col items-center justify-center">
+      <h1 className="font-roboto font-medium text-primary">Welcome to the start</h1>
+      <h1 className="font-poppins font-semibold text-primary">Welcome to the start</h1>
+      <h1 className="font-fira-sans font-bold text-primary">Welcome to the start</h1>
     </main>
   );
 }
